enhance: move GCP out of s3_filesystem_producer into its own producer
s3_filesystem_producer.cpp had accumulated a fair bit of GCP-specific code — GoogleHttpClientFactory, OAuth2 Bearer injection for IAM mode, GOOG4 re-signing for conditional writes — all tangled up with the AWS/Aliyun/Tencent/Huawei paths. This pulls everything GCP out into a new GcpFileSystemProducer under cpp/src/filesystem/gcp/ that inherits straight from FileSystemProducer. No shared base class or virtual hooks with the S3 side, the two producers are fully independent now.

The one thing both sides still share is TlsCurlHttpClient plus its TlsVersionToCurlOpt helper, which I pulled up to cpp/include/milvus-storage/filesystem/tls_http_client.h. It's a generic curl+AWS-SDK TLS knob, not really S3- or GCP-specific. TlsHttpClientFactory stays inside s3_filesystem_producer.cpp since only the S3 side uses it.

Worth flagging, this split relies on only one cloud provider being active per process, because InitializeS3 installs the HttpClientFactory globally via call_once. The old code had the exact same constraint, it just wasn't made explicit anywhere. If we ever need to mix providers in a single process we'll have to revisit this, but the split doesn't make the situation any worse.
